Fixing coverage issue on @plumier/core
